[Updated] Schematic listing method in datapad [Fixed] Food tool
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@2465 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/objects/resource/ResourceSpawnImplementation.cpp
@@ -236,11 +236,11 @@ ResourceContainer* ResourceSpawnImplementation::createResource(int units) {
 }
 
 void ResourceSpawnImplementation::decreaseContainerReferenceCount() {
-	/*if (--containerReferenceCount < 1 && !inShift()) {
+	if (--containerReferenceCount < 1 && !inShift()) {
 		destroyObjectFromDatabase(true);
 
 		dbDestroyed = true;
-	}*/
+	}
 }
 
 void ResourceSpawnImplementation::print() {
